x86/htirq: ipipe: enable interrupt pipelining
The Hypertransport handlers for masking/unmasking IRQs are invoked by
the pipeline core for dealing with incoming PCI interrupts from an
out-of-band context.

To this end, we have to turn the original spinlock into a hard
pipeline lock for proper serialization.

In addition, make sure to fix up the HT irq_chip descriptor
appropriately.