atomic: ipipe: keep atomic when pipelining IRQs
Because of the virtualization of interrupt masking for the regular
kernel code when the pipeline is enabled, atomic helpers relying on
common interrupt disabling helpers such as local_irq_save/restore
pairs would not be atomic anymore, leading to data corruption.

This commit restores true atomicity for the atomic helpers that would
be otherwise affected by interrupt virtualization.