context_tracking: ipipe: do not track over the head domain
Context tracking is a prerequisite for FULL_NOHZ, so that the RCU
subsystem can detect CPU idleness without relying on the (regular)
timer tick.

Out-of-band activity running over the head domain should by definition
not be involved in such detection logic, as the root domain has no
knowledge of what happens - and when - on the head domain whatsoever.