x86: ipipe: Do not dump panic traces outside of an oops
If the reported error leads to an oops, those messages can easily get
lost as they will be dumped only after the oops report - if we get
there and do not lock up earlier. And the dump by the oops will not find
a trace anymore as the first one consumed it. It is more important to
have the trace reliably on oopses than on warnings.

Move up the freeze a bit at this chance.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>