ipipe: mm: Drop un-COW from copy_pte_range
This is not needed, in fact, and it is in conflict with upcoming stable
changes. Philippe writes:

"IIRC, the reason was to ensure that any mapping duplicated from a
parent process which has pinned its memory via an earlier call to
__ipipe_disable_ondemand_mappings() would be immediately un-COWed upon
fork() in the child process. On second thoughts, this is useless since
the child process either:

- if it has to, explicitly attaches to the real-time core, which ends up
  calling mlock(), therefore causing these mappings to be un-COWed if
  need be. IOW, there is no inheritance of the Cobalt context at
  fork. Maybe at some point in time there was the assumption that such
  context might be shared between both processes somehow, which has been
  ruled out later on.

- the child process does not attach to the real-time core, in which case
  COW is a non-issue."

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>