x86: ipipe: Fix imbalance in slow syscall return path
When ipipe_handle_syscall signals "do not pass, return via slow path",
we must not run syscall_slow_exit_work. So far, this case was detected
by checking the syscall number. However, this missed the case that an
invalid syscall was passed down, forwarded by ipipe_handle_syscall, but
then ignored on exit due to the range check.

Instead, we need to pass the condition from the call site to make the
exit reliably balanced.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>